257 hipparcos spectral type (#258)
* #257: star database (Star.bin) now contains spectral type information (in the form of an integer index from 0 (O0) to 69 (M9)). Orbiter inline graphics and clients read the information and convert to colour variation for the pixel rendering of stars on the celestial sphere.

* #257: adjusted default star rendering paramters for slight increase in brightness.

* #257: API: replaced GraphicsClient::LoadStars with GraphicsClient::LoadStarData. Cleaned up the code of the LoadStars methods of the inline, D3D7 and D3D9 clients.

* #257: introduced a common method GraphicsClient::StarData2RenderData to process star data and map them to the celestial sphere. Celestial sphere is assumed to have radius 1 for rendering purposes. Camera near and farplanes are reset for celestial sphere rendering (inline, D3D7 and D3D9 clients)

* #257: unified loading and processing of constellation lines. Lines are now projected onto celestial sphere with radius 1.

* #257: converted constellation line vertices to vertex buffer (inline, D3D7, D3D9).

* #257: vertices for grid lines are now mapped to the celestial unit sphere (radius 1)

* #257: Updated inline client Scene::LoadStars(). Fixed bug in GraphicsClient::LoadStarData(). Converted array of star vertex buffers to std::vector (inline, D3D7, D3D9). Modified the colour curves for spectral class mapping (less blue).

* #257: added CelestialSphere class for inline client to make consistent with across clients.

* #257: inline graphics: moved RenderGrids function from Scene to CelestialSphere.

* #257: constellation labels: loading and preprocessing data is now done via GraphicsClient helper functions. Label rendering is now done by CelestialSphere implementations.

* #257: Switched LABELSPEC labels from char* to std::string.

* #257: LABELLIST: converted marker array to std::vector.

* #257: LABELLIST: converted name from char[64] to std::string.

* #257: Added class CelestialSphere to API to perform graphics-engine independent tasks (data reading and preprocessing) for celestial sphere.

* #257: Moved loading and preprocessing of constellation line data to base class CelestialSphere.

* #257: Replaced Scene::RenderDirectionMarker with CelestialSphere::RenderMarker. Replaced HDC-based label rendering with Sketchpad. Cleaned up marker drawing on celestial sphere.

* #257: Unified rendering of celestial markers between clients. Re-activated celestial background image rendering for inline client.

* #257: Inine graphics: moved celestial sphere background image management from Scene to CelestialSphere.

* #257: D3D7Client: moved celestial sphere background image management from Scene to CelestialSphere.

* #257: D3D9Client: moved celestial sphere background image management from Scene to CelestialSphere.

* #257: D3D9Client: use generic CelestialSphere method for rendering star markers.

* #257: D3D9Client: replaced local Sketchpad resources with GraphicsClient calls.

* #257: Inline graphics: Rendering of celestial sphere moved from Scene to CelestialSphere.

* #257: D3D7Client: celestial sphere rendering has been moved from Scene to CelestialSphere class.

* #257: D3D7Client: fixed black sky and z-tearing problems.

* #257: D3D9Client: moved celestial sphere rendering from Scene to CelestialSphere.

* #257: CelestialSphere: base class now provides methods to adjust marker and text colours against a bright sky background. Computation of star cutoff indices has been moved from individual clients to CelestialSphere base class.

* #257: All graphics clients: removed arbitrary radius (2000) from mesh used for celestial background image rendering. This caused the image not to be shown as the frustum far plane for rendering the celestial sphere is set to 10. D3D9Client: fixed wrong format for background brightness passed to celestial sphere background image renderer (was integer 0-255, is now float 0-1).

* #257: D3D9Client: 'target equator' planetarium feature is now implemented.

* #257: Updated star database to include van Leeuwen New Reduction. Changed colour rescaling to preserve colours for bright stars. Updated credits to include star database references.

* #257: Added Kuddel's updates to the DrawOrbits D3D9Client sample codes.

Closes #257.

+++ OVP/D3D9Client/CSphereMgr.h
@@ -48,7 +48,7 @@ public:
 	static void CreateDeviceObjects(LPDIRECT3D9 d3d, LPDIRECT3DDEVICE9 dev);
 	static void DestroyDeviceObjects();
 
-	void Render (LPDIRECT3DDEVICE9 dev, int level, int bglvl);
+	void Render (LPDIRECT3DDEVICE9 dev, int level, double bglvl);
 
 protected:
 	bool LoadPatchData ();
